move page type components
