+++ client/src/components/App.js
@@ -2,8 +2,8 @@ import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 
 import GlobalStyles from "./styles/GlobalStyles";
 import Header from "./Header";
-import Home from "./Home";
-import SignIn from "./SignIn";
+import Home from "../pages/Home";
+import SignIn from "../pages/SignIn";
 import SignUp from "../pages/SignUp";
 
 import { STRINGS } from "../constants";
